Moved simulation to top level
